+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.1] - 2026-06-23
+
+### Changed
+
+- windows update fix
+- changed `/sandbox none` to `/dandbox off` for consistency
+- changed /delete command to use python slice format
+- added cwd to system prompt and fixed local save names
+
 ## [0.2.0] - 2026-06-22
 
 ### Added